+++ checkout-ui-custom/src/_js/_customAddressForm.js
@@ -42,6 +42,8 @@ class fnsCustomAddressForm {
     this.validate = true
 
     this.gPlacesAutocomplete = ''
+
+    this.firstAttempt=false;
   }
 
   loadScript() {
@@ -384,74 +386,103 @@ class fnsCustomAddressForm {
 
     $('body').addClass('js-v-custom-is-loading')
 
-    fetch(
-      `/api/checkout/pub/orderForm/${_this.orderForm.orderFormId}/attachments/shippingData`,
-      {
-        credentials: 'include',
-        headers: {
-          accept: 'application/json, text/javascript, */*; q=0.01',
-          'cache-control': 'no-cache',
-          'content-type': 'application/json; charset=UTF-8',
-          pragma: 'no-cache',
-          'sec-fetch-mode': 'cors',
-          'sec-fetch-site': 'same-origin',
-          'x-requested-with': 'XMLHttpRequest',
-        },
-        referrerPolicy: 'no-referrer-when-downgrade',
-        body: JSON.stringify({
-          selectedAddresses: [
-            {
-              addressType: 'residential',
-              receiverName: '',
-              addressId: '',
-              isDisposable: true,
-              postalCode: _postalCode,
-              city: _city,
-              state: _state,
-              country: _country,
-              geoCoordinates,
-              street: _street,
-              number: _number || '',
-              neighborhood: _neighborhood,
-              complement: _complement,
-              reference: null,
-              addressQuery: _addressQuery,
-            },
-          ],
-          clearAddressIfPostalCodeNotFound: false,
-        }),
-        method: 'POST',
-        mode: 'cors',
-      }
-    )
-      .then(response => response.json())
-      .then(function (data) {
-        if (data.error) {
+    // fetch(
+    //   `/api/checkout/pub/orderForm/${_this.orderForm.orderFormId}/attachments/shippingData`,
+    //   {
+    //     credentials: 'include',
+    //     headers: {
+    //       accept: 'application/json, text/javascript, */*; q=0.01',
+    //       'cache-control': 'no-cache',
+    //       'content-type': 'application/json; charset=UTF-8',
+    //       pragma: 'no-cache',
+    //       'sec-fetch-mode': 'cors',
+    //       'sec-fetch-site': 'same-origin',
+    //       'x-requested-with': 'XMLHttpRequest',
+    //     },
+    //     referrerPolicy: 'no-referrer-when-downgrade',
+    //     body: JSON.stringify({
+    //       selectedAddresses: [
+    //         {
+    //           addressType: 'residential',
+    //           receiverName: '',
+    //           addressId: '',
+    //           isDisposable: true,
+    //           postalCode: _postalCode,
+    //           city: _city,
+    //           state: _state,
+    //           country: _country,
+    //           geoCoordinates,
+    //           street: _street,
+    //           number: _number || '',
+    //           neighborhood: _neighborhood,
+    //           complement: _complement,
+    //           reference: null,
+    //           addressQuery: _addressQuery,
+    //         },
+    //       ],
+    //       clearAddressIfPostalCodeNotFound: false,
+    //     }),
+    //     method: 'POST',
+    //     mode: 'cors',
+    //   }
+    // )
+
+    const shippingInfo = {
+      "selectedAddresses": [{
+            "addressType": "residential",
+            "receiverName": "",
+            "isDisposable": true,
+            "postalCode": _postalCode,
+            "city": _city,
+            "state": _state,
+            "country": _country,
+            "street":_street,
+            "number": _number || '',
+            "neighborhood": _neighborhood,
+            "complement": _complement,
+            "reference": null,
+            "geoCoordinates": geoCoordinates || [],
+            "addressQuery": _addressQuery
+        }],
+      "clearAddressIfPostalCodeNotFound": false
+    }
+
+    vtexjs.checkout.sendAttachment("shippingData", {}).done(function(orderForm) {
+      $("button.vtex-front-messages-close-all.close").trigger("click");
+      $(".vtex-omnishipping-1-x-warning").hide();
+      _this.firstAttempt = true;
+      vtexjs.checkout.sendAttachment("shippingData", shippingInfo).done(function(orderForm) {
+        if (orderForm.error) {
           $('body').removeClass('js-v-custom-is-loading')
           // eslint-disable-next-line no-alert
-          alert(`Something went wrong: ${data.error.message}`)
+          alert(`Something went wrong: ${orderForm.error.message}`)
         } else {
-          window.vtexjs.checkout.getOrderForm().done(function () {
-            _this.updateAddress(
-              _country,
-              _postalCode,
-              _city,
-              _state,
-              _street,
-              _number,
-              _complement,
-              '',
-              _addressQuery || '',
-              _addressId || '',
-              geoCoordinates || []
-            )
-            $('body').removeClass(_this.BodyFormClasses.join(' '))
-            _this.orderForm = window.vtexjs.checkout.orderForm
-            $('body').removeClass('js-v-custom-is-loading')
-            _this.triggerAddressValidation()
-          })
+          _this.updateAddress(
+            _country,
+            _postalCode,
+            _city,
+            _state,
+            _street,
+            _number,
+            _complement,
+            '',
+            _addressQuery || '',
+            _addressId || '',
+            geoCoordinates || []
+          )
+          $('body').removeClass(_this.BodyFormClasses.join(' '))
+          _this.orderForm = window.vtexjs.checkout.orderForm
+          $('body').removeClass('js-v-custom-is-loading')
+          _this.triggerAddressValidation()
         }
       })
+      .fail(function(error) {
+        $('body').removeClass(_this.BodyFormClasses.join(' '))
+        _this.orderForm = window.vtexjs.checkout.orderForm
+        $('body').removeClass('js-v-custom-is-loading')
+        alert(`Something went wrong: ${error}`)
+      })
+    })
   }
 
   getRegions(country) {
@@ -655,9 +686,9 @@ class fnsCustomAddressForm {
       $('.orderform-template-holder #shipping-data').append(form)
     }
 
-    if (orderForm.canEditData) {
-      $('body').removeClass('v-custom-addressForm-on')
-    }
+    // if (orderForm.canEditData) {
+    //   $('body').removeClass('v-custom-addressForm-on')
+    // }
 
     if (
       $('#shipping-option-pickup-in-point').hasClass(
@@ -669,7 +700,7 @@ class fnsCustomAddressForm {
       !$('body').hasClass('v-custom-addressForm-on') &&
       shippingData.selectedAddresses.length == 0
     ) {
-      $('body').addClass('v-custom-addressForm-on')
+      //$('body').addClass('v-custom-addressForm-on')
     }
 
     this.googleForm()
@@ -936,6 +967,7 @@ class fnsCustomAddressForm {
       if (
         (orderForm.shippingData.address === null ||
           orderForm.shippingData.address.addressType === 'search') &&
+        (!_this.firstAttempt) &&
         $('.vtex-omnishipping-1-x-deliveryOptionActive').attr('id') ===
           'shipping-option-delivery'
       ) {
@@ -959,11 +991,21 @@ class fnsCustomAddressForm {
       })
   }
 
+  loadingAddress(orderForm) {
+
+    if(~window.location.hash.indexOf("#/shipping") &&  orderForm.shippingData.availableAddresses.length && orderForm.shippingData.address==null) {
+      $('body').addClass('js-v-custom-is-loadAddress')
+    } else {
+      $('body').removeClass('js-v-custom-is-loadAddress')
+    }
+  }
+
   events() {
     const _this = this
 
     $(window).on('orderFormUpdated.vtex', function (evt, orderForm) {
       _this.checkFirstLogin(orderForm)
+      _this.loadingAddress(orderForm)
     })
   }
 
@@ -976,6 +1018,7 @@ class fnsCustomAddressForm {
       $('.vcustom--vtex-omnishipping-1-x-address').length < 1 &&
       orderForm.items.length
     ) {
+
       $('body').addClass(`${this.classOn}`)
       _this.orderForm = orderForm
       _this