+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+### Fixed
+
+- Major change in the google address form function
+
 ## [0.10.1] - 2023-04-27
 
 ### Fixed